Method Name fix, toString Poloymorphism, added Getter
Upon re-reading specification over, changed the name of the method that was originally named "getNumOfGrains" to "getGenerationCount". Overrode the toString method inherited from item class to treat a Grain object as either "g" or "G" depending on it age compared to its maturation age. Created a getter for the "grain cost" constant.