Indentation and farewell message placement
Tested some gameplay, ran into the issue of the farewell message being printed everytime any action was made. This was because the message was inside the main loop, which occurs while the game is running. Thus, the message would appear at every iteration instead of when the player terminates the loop via "q".